updating our project

+++ store/templatetags/cart.py
@@ -2,6 +2,7 @@ from django import template
 
 register = template.Library()
 
+
 @register.filter(name='is_in_cart')
 def is_in_cart(product  , cart):
     keys = cart.keys()
@@ -12,24 +13,23 @@ def is_in_cart(product  , cart):
 
 
 @register.filter(name='cart_quantity')
-def cart_quantity(product  , cart):
+def cart_quantity(product, cart):
     keys = cart.keys()
     for id in keys:
-        if int(id) == product.id:
+        if float(id) == product.id:
             return cart.get(id)
     return 0;
 
 
 @register.filter(name='price_total')
-def price_total(product  , cart):
-    return product.price * cart_quantity(product , cart)
+def price_total(product, cart):
+    return product.price * cart_quantity(product, cart)
 
 
 @register.filter(name='total_cart_price')
-def total_cart_price(products , cart):
-    sum = 0 ;
+def total_cart_price(products, cart):
+    sum = 0;
     for p in products:
-        sum += price_total(p , cart)
+        sum += price_total(p, cart)
 
     return sum
-
